Refactor play data upload to backend
